+++ src/main/java/com/jnape/palatable/lambda/functions/specialized/Predicate.java
@@ -37,8 +37,7 @@ public interface Predicate<A> extends Fn1<A, Boolean>, java.util.function.Predic
      * {@inheritDoc}
      */
     @Override
-    @SuppressWarnings("unchecked")
-    default <Z> Predicate<Z> diMapL(Function<Z, A> fn) {
+    default <Z> Predicate<Z> diMapL(Function<? super Z, ? extends A> fn) {
         return Fn1.super.diMapL(fn)::apply;
     }