+++ src/main/java/com/jnape/palatable/lambda/functor/Profunctor.java
@@ -31,7 +31,7 @@ public interface Profunctor<A, B, PF extends Profunctor> {
      * @param fn  the mapping function
      * @return a profunctor over Z (the new left parameter type) and C (the same right parameter type)
      */
-    default <Z> Profunctor<Z, B, PF> diMapL(Function<Z, A> fn) {
+    default <Z> Profunctor<Z, B, PF> diMapL(Function<? super Z, ? extends A> fn) {
         return diMap(fn, id());
     }
 
@@ -43,7 +43,7 @@ public interface Profunctor<A, B, PF extends Profunctor> {
      * @param fn  the mapping function
      * @return a profunctor over A (the same left parameter type) and C (the new right parameter type)
      */
-    default <C> Profunctor<A, C, PF> diMapR(Function<B, C> fn) {
+    default <C> Profunctor<A, C, PF> diMapR(Function<? super B, ? extends C> fn) {
         return diMap(id(), fn);
     }
 
@@ -57,5 +57,5 @@ public interface Profunctor<A, B, PF extends Profunctor> {
      * @param rFn the right parameter mapping function
      * @return a profunctor over Z (the new left parameter type) and C (the new right parameter type)
      */
-    <Z, C> Profunctor<Z, C, PF> diMap(Function<Z, A> lFn, Function<B, C> rFn);
+    <Z, C> Profunctor<Z, C, PF> diMap(Function<? super Z, ? extends A> lFn, Function<? super B, ? extends C> rFn);
 }